PlayerController Interactable Mesh Collider + @GameController
Just made a comment by the code in Awake() to create a Sphere collider if interactionDetector is null; just change this to create a MeshCollider if null?

Also renamed GameController to @GameController for consistency; changed the name reference in scripts that needed it (i.e. Billboard UI and Button Prompt)