Headbuttable Objects self-direct to activate Interaction when Headbutted
Adding to the list first was actually negating the Interaction because the headbuttable object had not yet been added to the list.

Bypassing the list and simply telling the headbuttedobject to activate its Interact() in OnTriggerEnter seems to have eliminated the problem and made a headbuttables list redundant.

Commented out the list and its references in playerController for now and otherwise left them untouched, just in case a need for it arises later.