Basic Area Title Card
Canvas and Player Detector prefabs are available in Prefabs > Systems > UI > World UI > AreaTitle.

Canvas is already hooked up to the Player Detector and does not need to be placed in the scene. Implementer need only type what they want the title card to say in the Player Detector' Area Title field and place it where the player will contact it. The canvas then instantiates, plays, and deletes itself.