Update
Starting new Player_v4.0 Prefab with colliders to start helping with infinite sliding.
Had to add the tag 'Player' to them to start getting them working with pickups and then had to edit the following scripts to search for playercontroller in their parents because the colliders are no longer directly on the same object as the PlayerController.